# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ApunBazar E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filtering
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, popular]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist by session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add item to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToWishlistBody"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove item from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        featured:
          type: boolean
        artisan:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, name, description, price, categoryId, categoryName, categorySlug, imageUrl, images, stock, rating, reviewCount, featured, tags, createdAt]
    ProductsListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, limit, totalPages]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        artisan:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [name, description, price, categoryId, imageUrl, stock]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        artisan:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
...
    WishlistItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        sessionId:
          type: string
        addedAt:
          type: string
      required: [id, productId, product, sessionId, addedAt]
    AddToWishlistBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
      required: [sessionId, productId]
...